backend/requirements.txt
fastapi
uvicorn
supabase
openai
aiohttp
python-dotenv
python-multipart
pydantic